+++ tools/exportPgn/tool.js
@@ -10,6 +10,14 @@
         defaultValue: true,
         advanced: true,
         hidden: true
+      },
+      {
+        name: 'exportPGNoptions',
+        category: 'analysis',
+        type: 'multiple',
+        possibleValues: ['exportClock', 'exportEval'],
+        defaultValue: 'exportClock,exportEval',
+        advanced: true
       }
     ];
 
@@ -18,13 +26,19 @@
         'options.analysis': 'Analysis',
         'options.exportPGN': 'Export PGN',
         'PGNCopiedToClipboard': 'PGN copied to clipboard',
-        'clipboardDenied': 'Clipboard access denied'
+        'clipboardDenied': 'Clipboard access denied',
+        'options.exportPGNoptions': 'Options for PGN exports',
+        'exportPGNoptions.exportClock': 'Export clock values',
+        'exportPGNoptions.exportEval': 'Export computer evaluation'
       },
       'ro-RO': {
         'options.analysis': 'Analiz\u0103',
         'options.exportPGN': 'Export\u0103 PGN',
         'PGNCopiedToClipboard': 'PGN copiat \u00een clipboard',
-        'clipboardDenied': 'Acces refuzat la clipboard'
+        'clipboardDenied': 'Acces refuzat la clipboard',
+        'options.exportPGNoptions': 'Op\u0163iuni pentru exporturi PGN',
+        'exportPGNoptions.exportClock': 'Export\u0103 timp pe mut\u0103ri',
+        'exportPGNoptions.exportEval': 'Export\u0103 evalu\u0103ri computer'
       }
     }
 
@@ -35,6 +49,8 @@
         toPosition: false,
         separateLines: false,
         unicode: false,
+        exportClock: this.options.exportClock,
+        exportEval: this.options.exportEval,
         ...options
       };
       const lt = this.lichessTools;
@@ -94,7 +110,7 @@
           const code = shape.brush[0].toUpperCase() + shape.orig + (shape.dest || '');
           group.shapes.push(code);
         }
-        if (node.clock) {
+        if (options.exportClock && node.clock) {
           const group = {
             type: 'clk',
             shapes: [ centisecondsToClock(node.clock) ]
@@ -102,7 +118,7 @@
           groups.push(group);
         }
         const evl = node.ceval || node.eval;
-        if (evl) {
+        if (options.exportEval && evl) {
           const group = {
             type: 'eval',
             shapes: [ evalToString(evl) ]
@@ -284,7 +300,14 @@
     async start() {
       const lt = this.lichessTools;
       const value = lt.currentOptions.getValue('exportPGN');
+      const opts = lt.currentOptions.getValue('exportPGNoptions');
       this.logOption('Export PGN', value);
+      this.logOption('Export PGN options', opts);
+      this.options = {
+        exportClock: lt.isOptionSet(opts, 'exportClock'),
+        exportEval: lt.isOptionSet(opts, 'exportEval')
+      };
+
       if (value) {
         lt.exportPgn = this.exportPgn;
       } else {