+++ HISTORY.md
@@ -3,6 +3,7 @@ History of features added to LiChess Tools in time (reversed order)
 07 Jan
 
 - puzzle session total
+- PGN export options
 
 30 Dec